automatically generate test cases when tests are run
