+++ scripts/prebuild.js
